Goodbye, GraphQL
