--- app/client/graphQL/constants.ts
@@ -1,16 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-export const QUERY_ENTRY = 'gql_m_entry';
-export const QUERY_CHANNELS = 'gql_m_channels';
-export const QUERY_CHANNELS_NEXT = 'gql_m_channels_next';
-export const QUERY_ALL_CHANNELS = 'gql_m_all_channels';
-export const QUERY_ALL_CHANNELS_NEXT = 'gql_m_all_channels_next';
-
-export default {
-    QUERY_ENTRY,
-    QUERY_CHANNELS,
-    QUERY_CHANNELS_NEXT,
-    QUERY_ALL_CHANNELS,
-    QUERY_ALL_CHANNELS_NEXT,
-};
