--- app/constants/graphql.ts
@@ -1,4 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-export const MEMBERS_PER_PAGE = 200;
